Introduce node-insensitive Value compare & search
Add functions for testing equality of only the "BASIC" parts
of two StellarValues, which means ignoring the node ID and
signature, and a function for searching a given vector of
StellarValues for one which matches a given StellarValue
under the new node-insensitive equality.

If the new node-insensitive equality test is given one or
more `Value`s which are not opaque `StellarValue`s, it tests
the parameters for strict equality.  In production, we
should never be comparing non-`StellarValue`s with the new
function, but some tests induce votes for `Value`s which are
not `StellarValue`s.