Treat votes from different nodes as equivalent
Experiment with whether the number of nomination messages
can be (correctly!) reduced by treating a vote for a value
as a vote for any value with equal BASIC components.  Since
the non-basic component is a node ID and signature, this
means that votes are treated as equivalent based on a
node-insensitive equality.

We continue to consider a value as _accepted_ by a node
only if that node has accepted that _exact_ value, including
which node nominated it.  This ensures that nodes that form
a quorum will agree on the exact list of candidates.