Map ClaimableBalanceID and LedgerKey to one byte
Compact `ClaimableBalanceID`s and `LedgerKey`s into a
single byte, and use that byte to index into an array of
IDs generated during initialization and setup, in the hope
of reducing the space that the fuzzer has to explore in
order to perform operations on sponsorships and claimable
balances created during setup.  In particular, it shouldn't
have to guess a hash anymore.

This comprises the following changes:

- When performing operations that can make `LedgerKey`s (or
  `ClaimableBalanceID`s) available for reference by future
  operations, store those `LedgerKey`s in the byte-indexed
  array.
- In the XDR compactor, convert `LedgerKey` and
  `ClaimableBalanceID` into a single byte (the first
  byte of the opaque XDR representation).
- In the XDR unpacker, unpack `LedgerKey`s and
  `ClaimableBalanceID`s into ones from the stored array of
  `LedgerKey`s.
- Create some default pseudorandom `LedgerKey`s when
  creating a `Fuzzer`.  (Some of them will be overwritten
  by new ones generated during setup.)