Loop to generate valid fuzz corpus transactions
We're postulating that it's easier for the fuzzer to find
invalid transactions given valid ones in the corpus than the
other way around.

Checking operation validity in TransactionFuzzer::genFuzz()
requires us to create a FuzzTransactionFrame, and therefore
to have an Application and a root TestAccount.  So we factor
that process out of TransactionFuzzer::initialize() into a
method and call it from genFuzz() as well.